Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/html/semantics/scripting-1/the-script-element/module/specifier-error-expected.txt

CONSOLE MESSAGE: TypeError: Module specifier, 'string-without-dot-slash-prefix' does not start with "/", "./", or "../". Referenced from http://localhost:8800/html/semantics/scripting-1/the-script-element/module/bad-module-specifier.js

FAIL Test that invalid module specifier leads to TypeError on window. assert_equals: expected function "function TypeError() {
    [native code]
}" but got function "function String() {
    [native code]
}"
